requirements.txt
altgraph==0.17.3
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
numpy==1.25.1
oauthlib==3.2.2
pandas==2.0.3
pefile==2023.2.7
py-trello==0.19.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
XlsxWriter==3.1.2
